+++ src/main.rs
@@ -48,15 +48,6 @@ fn main() -> amethyst::Result<()> {
             systems::ActionTrackerSystem,
             "action_tracker",
             &["input_system"],
-        )
-        .with(
-            systems::BillboardDisplaySystem {
-                glyph_speed: std::env::var("TALKIE_SPEED")
-                    .map(|s| s.parse().expect("invalid speed."))
-                    .ok(),
-            },
-            "billboard_display",
-            &["dialogue_processor"],
         );
 
     let mut game = Application::new(
